requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
dateparser==1.2.0
Flask==3.0.2
Flask-Cors==4.0.0
frozenlist==1.5.0
gunicorn==21.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mkl-service==2.5.2
multidict==6.1.0
numpy==1.26.4
bottleneck==1.3.8
pandas==2.2.2
cython==3.0.0
propcache==0.2.0
pycryptodome==3.21.0
python-binance==1.0.22
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.11.6
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
Werkzeug==3.0.1
websockets==14.1
yarl==1.18.0
python-dotenv==1.0.0
pymongo==4.7.2